Some fixes

+++ database/migrations/2014_04_05_143306_migrations_create_page.php
@@ -22,7 +22,7 @@ class MigrationsCreatePage extends Migration {
                    $table->tinyInteger('in_list')->default(1)->index();
                    $table->tinyInteger('in_last')->default(1)->index();
                    $table->tinyInteger('in_news')->default(1)->index();
-                   $table->tinyInteger('manual_order')->default(0);
+                   $table->bigInteger('manual_order')->default(0);
                    $table->tinyInteger('views')->default(0);
                    $table->tinyInteger('hidden')->default(0)->index();
                    $table->nullableTimestamps();
